chore: финальные правки
