+++ package.json
@@ -11,7 +11,6 @@
     },
     "dependencies": {
         "@editorjs/editorjs": "^2.30.8",
-        "@editorjs/list": "^2.0.8",
         "@editorjs/paragraph": "^2.11.7",
         "@editorjs/underline": "^1.2.1",
         "@hookform/resolvers": "^5.1.1",
