+++ src/components/ui/TextEditor/editorjs.config.ts
@@ -1,5 +1,4 @@
 import type { ToolConstructable, ToolSettings } from '@editorjs/editorjs/types/tools';
-import EditorjsList from '@editorjs/list';
 import Paragraph from '@editorjs/paragraph';
 import Underline from '@editorjs/underline';
 
@@ -14,21 +13,21 @@ export const BASE_EDITOR_CONFIG: {
             preserveBlank: true,
         },
     },
-    List: {
-        class: EditorjsList as unknown as ToolConstructable,
-        inlineToolbar: true,
-        config: {
-            defaultStyle: 'ordered',
-            maxLevel: 1,
-            counterTypes: ['numeric'],
-        },
-        toolbox: [
-            {
-                data: { style: 'unordered' },
-            },
-            {
-                data: { style: 'ordered' },
-            },
-        ],
-    },
+    // List: {
+    //     class: EditorjsList as unknown as ToolConstructable,
+    //     inlineToolbar: true,
+    //     config: {
+    //         defaultStyle: 'ordered',
+    //         maxLevel: 1,
+    //         counterTypes: ['numeric'],
+    //     },
+    //     toolbox: [
+    //         {
+    //             data: { style: 'unordered' },
+    //         },
+    //         {
+    //             data: { style: 'ordered' },
+    //         },
+    //     ],
+    // },
 };
